Divided tests into files
